booking cancellation & changed IDs to long

+++ src/main/java/itmo/tg/airbnb_business/business/dto/AdvertisementBlockDTO.java
@@ -13,7 +13,7 @@ import java.time.LocalDate;
 @Builder
 public class AdvertisementBlockDTO {
 
-    private Integer advertisementId;
+    private Long advertisementId;
 
     private LocalDate dateUntil;
