+++ src/main/java/itmo/tg/airbnb_business/business/dto/FineDTO.java
@@ -12,9 +12,9 @@ import lombok.NoArgsConstructor;
 @Builder
 public class FineDTO {
 
-    private Integer id;
+    private Long id;
 
-    private Integer amount;
+    private Double amount;
 
     private FineStatus status;
