+++ src/main/java/itmo/tg/airbnb_business/business/repository/FineRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FineRepository extends JpaRepository<Fine, Integer> {
+public interface FineRepository extends JpaRepository<Fine, Long> {
 
     List<Fine> findByUser(User user);
